+++ lib/pg/aws_rds_iam/active_record_postgresql_adapter.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module PG
+  module AWS_RDS_IAM
+    module ActiveRecordPostgreSQLAdapter
+      def dbconsole(config, *)
+        AuthTokenInjector.new.inject_into_psql_env! config.configuration_hash, ENV
+        super
+      end
+    end
+
+    private_constant :ActiveRecordPostgreSQLAdapter
+  end
+end