+++ test/acceptance/rails/application.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require "pg/aws_rds_iam"
+
+TestApplication = Class.new(Rails::Application)